init: do not re-send the API key
There is no need to re-send the API key at the init phase: this has
already been done when validating the new configuration.

Sending the API key will make the Healthbox sending a request to the
cloud servers. If the Healthbox doesn't have Internet access at that
time, the integration will refuse to start.